requirements.txt
Flask
mysql-connector-python
psycopg2-binary
gunicorn